abstraction of api searchGet logic. Made singular call for unpaged/wrongly paged detail deeplinks
