+++ src/helpers/typeHelper.js
@@ -1,9 +1,3 @@
-import VideoDetail from '@/components/results/detail/VideoDetail';
-import DocumentDetail from '@/components/results/detail/DocumentDetail';
-import AudioDetail from '@/components/results/detail/AudioDetail';
-import DirectoryDetail from '@/components/results/detail/DirectoryDetail';
-import ImageDetail from '@/components/results/detail/ImageDetail';
-
 export const fileTypes = [
   'text',
   'images',
